Remove nullable types and add field constraints
Nullable types have been removed from Brand, VehicleCategory, and Game in Vehicle.cs, as well as in Vehicle and ApplicationUser in MtfstCarProSettings.cs. In addition, a max length constraint was added to the Description field in MtfstCarProSettings.cs and a Type field was added to VehicleCategory.cs with a max length constraint.