Refactor Cars page in Motorfest ProSettings
This update cleans up and refactors the Cars.cshtml.cs file within the Motorfest/ProSettings section of the project. The changes include removal of an unused import, minor code syntax corrections, and optimization of the search functionality in the OnGetLoadProSettingsAsync method. The method searchTokens are transformed to lowercase for case insensitive comparison.