Add image liking feature in DatabaseMethodService
A new method has been added to the DatabaseMethodService which allows users to like and unlike images. Along with this, a new log event for all database methods has been added. If there is an issue with updating the count of likes on a specific image, the new log event will record it.