Add delete feature and style for author tools
A new feature has been added for the authors to delete their images, highlighted with a small styled icon for delete function. The changes also include renaming the ToggleLike() method to ToggleLikeAsync() and some visual tweaks for better user interaction. Styling for the "author-tools" class has been added to the Image.razor.css, providing a nice hover effect with an emphasis on the delete icon.